Replace div with mul by the inverse in geo_dist
Multiplication is an order of magnitude faster than division. In this case,
the divisor is constant, meaning that it can be replace with a
multiplication by the inverse. The result will be practicaly the same
(float arithmetic is not commutative, therefore may not be strictly equal).

This patch improves the performance of a query on a GeoPositionToNode is
improved by about 9% on an Intel i7-4600U.

Signed-off-by: Francis Giraldeau <francis.giraldeau@gmail.com>