Add more places we can walk
This commit adds several areas we can also walk over, or walk through. The biggest missing one was highway=crossing which allows to walk from one side of the road to the other.

Other importants ones is the ability to walk over a public_transport=platform (or the old: highway=platform) tagging.

This change may introduce islands on places where such platforms, if they are not connected to the rest of the network.